+++ packages/react/.storybook/preview.tsx
@@ -2,7 +2,7 @@ import { MantineProvider, MantineThemeOverride } from '@mantine/core';
 import { MockClient } from '@medplum/mock';
 import { MedplumProvider } from '@medplum/react-hooks';
 import { BrowserRouter } from 'react-router-dom';
-import sinon from 'sinon';
+import { createGlobalTimer } from '../src/stories/MockDateWrapper.utils';
 
 export const parameters = {
   layout: 'fullscreen',
@@ -16,11 +16,10 @@ export const parameters = {
   },
 };
 
-const clock = sinon.useFakeTimers({
-  now: new Date(2020, 4, 4, 3, 14),
-  shouldAdvanceTime: false,
-  toFake: ['Date'],
-});
+// wrap intialization of MockClient and initial page navigation
+// so that resources created in MockFetchClient#initMockRepo have
+// consistent timestampt between storybook runs
+const clock = createGlobalTimer();
 const medplum = new MockClient();
 medplum.get('/').then(() => {
   clock.restore();
