+++ packages/react/src/RequestGroupDisplay/RequestGroupDisplay.stories.tsx
@@ -13,7 +13,7 @@ import {
   Covid19ReviewLabsTask,
 } from '../stories/covid19';
 import { RequestGroupDisplay } from './RequestGroupDisplay';
-import { withMockedDate } from '../utils/MockDateWrapper';
+import { withMockedDate } from '../stories/decorators';
 
 export default {
   title: 'Medplum/RequestGroupDisplay',
