+++ packages/react/src/stories/decorators.tsx
@@ -1,0 +1,10 @@
+import { Decorator } from '@storybook/react';
+import { MockDateWrapper } from './MockDateWrapper';
+
+export const withMockedDate: Decorator = (Story) => {
+  return (
+    <MockDateWrapper>
+      <Story />
+    </MockDateWrapper>
+  );
+};
